+++ RELEASENOTES.md
@@ -77,6 +77,8 @@
     ([#4145](https://github.com/google/ExoPlayer/issues/4145)).
   * Preeptively declare an ID3 track in chunkless preparation
     ([#4016](https://github.com/google/ExoPlayer/issues/4016)).
+  * Add support for multiple #EXT-X-MAP tags in a media playlist
+    ([#4164](https://github.com/google/ExoPlayer/issues/4182)).
 * Fix ClearKey decryption error if the key contains a forward slash
   ([#4075](https://github.com/google/ExoPlayer/issues/4075)).
 * Fix crash when switching surface on Huawei P9 Lite